refactor: service layer to handle exceptions

+++ src/main/java/com/example/woowa/delivery/service/RiderService.java
@@ -12,6 +12,7 @@ import com.example.woowa.delivery.repository.AreaCodeRepository;
 import com.example.woowa.delivery.repository.RiderAreaCodeRepository;
 import com.example.woowa.delivery.repository.RiderRepository;
 import java.util.List;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -35,30 +36,49 @@ public class RiderService {
 
     @Transactional
     public void deleteAll() {
-        riderRepository.deleteAll();
+        try {
+            if (riderRepository.count() == 0) {
+                throw new RuntimeException(ErrorMessage.NOT_FOUND_DATA.getMessage());
+            }
+            riderRepository.deleteAll();
+        } catch (Exception e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_DELETE.getMessage());
+        }
     }
 
     @Transactional
     public Long save(RiderCreateRequest riderCreateRequest) {
-        Rider rider = riderMapper.toRider(riderCreateRequest);
-        boolean isId = riderRepository.existsByLoginId(riderCreateRequest.loginId());
-        if (isId) {
-            throw new RuntimeException(ErrorMessage.DUPLICATE_LOGIN_ID.getMessage());
-        }
+        try {
+            Rider rider = riderMapper.toRider(riderCreateRequest);
+            boolean isId = riderRepository.existsByLoginId(riderCreateRequest.loginId());
+            if (isId) {
+                throw new RuntimeException(ErrorMessage.DUPLICATE_LOGIN_ID.getMessage());
+            }
 
-        riderRepository.save(rider);
-        return rider.getId();
+            riderRepository.save(rider);
+            return rider.getId();
+        } catch (Exception e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_SAVE.getMessage());
+        }
     }
 
     @Transactional
     public void update(Long id, RiderUpdateRequest riderUpdateRequest) {
-        Rider rider = findEntityById(id);
-        rider.update(riderUpdateRequest.getName(), riderUpdateRequest.getPhoneNumber());
+        try {
+            Rider rider = findEntityById(id);
+            rider.update(riderUpdateRequest.getName(), riderUpdateRequest.getPhoneNumber());
+        } catch (RuntimeException e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_UPDATE.getMessage());
+        }
     }
 
     @Transactional
     public void changeIsDelivery(Long id, Boolean isDelivery) {
-        findEntityById(id).changeIsDelivery(isDelivery);
+        try {
+            findEntityById(id).changeIsDelivery(isDelivery);
+        } catch (RuntimeException e) {
+            throw new RuntimeException("배달 상태"+ErrorMessage.FAIL_TO_UPDATE.getMessage());
+        }
     }
 
     public RiderResponse findResponseById(Long id) {
@@ -70,34 +90,47 @@ public class RiderService {
     }
 
     public Page<RiderResponse> findAll(PageRequest pageRequest) {
-        return riderRepository.findAllBy(pageRequest).map(riderMapper::toResponse);
+        try {
+            return riderRepository.findAllBy(pageRequest).map(riderMapper::toResponse);
+        } catch (Exception e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_RETRIEVE.getMessage());
+        }
     }
 
     @Transactional
     public void addRiderAreaCode(Long riderId, Long areaCodeId) {
-        Rider rider = findEntityById(riderId);
-        AreaCode areaCode = areaCodeService.findEntityById(areaCodeId);
-        RiderAreaCode riderAreaCode = new RiderAreaCode(rider, areaCode);
-        riderAreaCodeRepository.save(riderAreaCode);
-        areaCode.addRiderAreaCode(riderAreaCode);
+        try {
+            Rider rider = findEntityById(riderId);
+            AreaCode areaCode = areaCodeService.findEntityById(areaCodeId);
+
+            RiderAreaCode riderAreaCode = new RiderAreaCode(rider, areaCode);
+            riderAreaCodeRepository.save(riderAreaCode);
+            areaCode.addRiderAreaCode(riderAreaCode);
+        } catch (Exception e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_SAVE.getMessage());
+        }
     }
 
     @Transactional
     public void removeRiderAreaCode(Long riderId, Long areaCodeId) {
-        Rider rider = findEntityById(riderId);
-        List<RiderAreaCode> riderAreaCodeList = rider.getRiderAreaCodeList();
-        riderAreaCodeList.stream()
-                .filter(riderAreaCode -> riderAreaCode.getAreaCode().getId().equals(areaCodeId))
-                .findFirst()
-                .ifPresent(riderAreaCode -> {
-                    rider.removeRiderAreaCode(riderAreaCode);
+        try {
+            Rider rider = findEntityById(riderId);
+            List<RiderAreaCode> riderAreaCodeList = rider.getRiderAreaCodeList();
+            riderAreaCodeList.stream()
+                    .filter(riderAreaCode -> riderAreaCode.getAreaCode().getId().equals(areaCodeId))
+                    .findFirst()
+                    .ifPresent(riderAreaCode -> {
+                        rider.removeRiderAreaCode(riderAreaCode);
 
-                    AreaCode areaCode=riderAreaCode.getAreaCode();
-                    areaCode.removeRiderAreaCode(riderAreaCode);
-                    areaCodeRepository.save(areaCode);
+                        AreaCode areaCode = riderAreaCode.getAreaCode();
+                        areaCode.removeRiderAreaCode(riderAreaCode);
+                        areaCodeRepository.save(areaCode);
 
-                    // 중간 엔티티 삭제
-                    riderAreaCodeRepository.delete(riderAreaCode);
-                });
+                        // 중간 엔티티 삭제
+                        riderAreaCodeRepository.delete(riderAreaCode);
+                    });
+        } catch (Exception e) {
+            throw new RuntimeException(ErrorMessage.FAIL_TO_DELETE.getMessage());
+        }
     }
 }
